All tests working wit 2 fails out of 9
