+++ check
@@ -1,1 +1,13 @@
 monitorEvents($0, 'mouse') //in chrome: to see the mouse events
+
+function monitorEvents(element) {
+  var log = function(e) { console.log(e);};
+  var events = [];
+
+  for(var i in element) {
+    if(i.startsWith("on")) events.push(i.substr(2));
+  }
+  events.forEach(function(eventName) {
+    element.addEventListener(eventName, log);
+  });
+}
